requirements.txt
asgiref==3.8.1
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
folium==0.19.5
idna==3.10
image==1.5.33
jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.3
pillow==11.1.0
pygments==2.19.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
xyzservices==2025.1.0
